+++ package.json
@@ -1,25 +1,21 @@
 {
   "name": "placemark-global",
   "dependencies": {
-    "@create-figma-plugin/ui": "^3.1.0",
-    "@create-figma-plugin/utilities": "^3.1.0",
-    "@types/d3": "^7.4.3",
-    "@types/d3-geo": "^3.1.0",
-    "@types/topojson-client": "^3.1.4",
     "d3": "^7.8.5",
     "d3-geo": "^3.1.0",
-    "preact": ">=10",
     "topojson-client": "^3.1.0"
   },
   "devDependencies": {
-    "@create-figma-plugin/build": "^3.1.0",
-    "@create-figma-plugin/tsconfig": "^3.1.0",
+    "@types/d3": "^7.4.3",
+    "@types/d3-geo": "^3.1.0",
+    "@types/topojson-client": "^3.1.4",
+    "vite": "^5.1.6",
     "@figma/plugin-typings": "1.82.0",
     "typescript": ">=5"
   },
   "scripts": {
-    "build": "build-figma-plugin --typecheck --minify",
-    "watch": "build-figma-plugin --typecheck --watch",
+    "dev": "vite",
+    "build": "vite build",
     "tsc": "tsc"
   },
   "figma-plugin": {
@@ -34,7 +30,7 @@
     },
     "name": "Placemark Globe",
     "main": "src/main.ts",
-    "ui": "src/ui.tsx"
+    "ui": "src/ui.ts"
   },
   "volta": {
     "node": "20.10.0"
