+++ src/index.html
@@ -1,0 +1,38 @@
+<html>
+  <style>
+    body {
+      font-family: "Inter", "Helvetica", sans-serif;
+      font-size: 12px;
+    }
+    select {
+      font: inherit;
+      width: 100%;
+      border: 1px solid #ccc;
+      border-radius: 0;
+    }
+  </style>
+  <script type="module" src="./ui.ts"></script>
+
+  <div style="padding: 10px">
+    <svg data-target="globe" style="height: 300px; width: 300px"></svg>
+    <div style="padding-top: 10px">
+      <label style="padding-bottom: 5px; display: block" for="dataset-selector"
+        >Dataset</label
+      >
+      <select id="dataset-selector" data-target="dataset-selector">
+        <option value="visionscarto">visionscarto</option>
+        <option value="naturalearth">naturalearth</option>
+      </select>
+    </div>
+    <div style="padding-top: 10px">
+      <a
+        style="color: blue"
+        target="blank"
+        rel="noreferrer"
+        href="https://gist.github.com/tmcw/22a083572f3ef478a64dce17680def08"
+      >
+        About data sources
+      </a>
+    </div>
+  </div>
+</html>
