+++ vite.config.js
@@ -1,0 +1,9 @@
+// vite.config.js
+export default {
+  // config options
+  root: "src",
+  build: {
+    watch: true,
+    outDir: "dist",
+  },
+};
